fix: build full-cache metadata for the replicated DCP indexer
Share the existing replicated address helpers through common_cp while keeping indexer metadata and buffers independent from SFA. Preserve upstream block-table interfaces and cache allocation.
